use gradient function
